Lefan abstraction work in the first week of September

+++ regression/abstraction/test1/main.c
@@ -7,9 +7,9 @@
 bool main(){
     char * a1;
     char * a2;
-    int a1_len;
-    int a2_len;
-    int i = 0;
+    size_t a1_len;
+    size_t a2_len;
+    size_t i = 0;
     bool res = true;
 
     __CPROVER_assume(a1_len < MAX_LEN);
