+++ regression/abstraction/test7/main.c
@@ -6,9 +6,9 @@
 void main(){
     char * a1;
     char * a2;
-    int a1_len;
-    int a2_len;
-    int i = 0;
+    size_t a1_len;
+    size_t a2_len;
+    size_t i = 0;
     int sum;
 
     __CPROVER_assume(a1_len < MAX_LEN);
